Added staking/slashes formula and validator formula types.
