+++ src/data/formulas/index.ts
@@ -1,3 +1,4 @@
 export * as contractFormulas from './contract'
-export * as walletFormulas from './wallet'
 export * as genericFormulas from './generic'
+export * as validatorFormulas from './validator'
+export * as walletFormulas from './wallet'
