+++ src/data/utils.ts
@@ -4,100 +4,51 @@ import {
   GenericFormula,
   NestedFormulaMap,
   TypedFormula,
+  ValidatorFormula,
   WalletFormula,
 } from '@/core'
 
-import { contractFormulas, genericFormulas, walletFormulas } from './formulas'
+import {
+  contractFormulas,
+  genericFormulas,
+  validatorFormulas,
+  walletFormulas,
+} from './formulas'
 
-export const getContractFormula = (
-  formulaName: string
-): ContractFormula<any, any> | undefined => {
-  const formulaPath = formulaName.split('/')
-  const formulaBase = formulaPath
-    .slice(0, -1)
-    .reduce(
-      (acc, key) =>
-        typeof acc === 'object' && key in acc
-          ? (acc as NestedFormulaMap<ContractFormula<any, any>>)[key]
-          : undefined,
-      contractFormulas as
-        | NestedFormulaMap<ContractFormula<any, any>>
-        | ContractFormula<any, any>
-        | undefined
-    )
+export const makeGetFormula =
+  <T extends unknown>(formulas: NestedFormulaMap<T> | T | undefined) =>
+  (formulaName: string): T | undefined => {
+    const formulaPath = formulaName.split('/')
+    const formulaBase = formulaPath
+      .slice(0, -1)
+      .reduce(
+        (acc, key) =>
+          acc && typeof acc === 'object' && key in acc
+            ? (acc as NestedFormulaMap<T>)[key]
+            : undefined,
+        formulas
+      )
 
-  const formula =
-    typeof formulaBase === 'object'
-      ? (formulaBase as NestedFormulaMap<ContractFormula<any, any>>)[
-          formulaPath[formulaPath.length - 1]
-        ]
+    const formula =
+      typeof formulaBase === 'object'
+        ? (formulaBase as NestedFormulaMap<ContractFormula<any, any>>)[
+            formulaPath[formulaPath.length - 1]
+          ]
+        : undefined
+
+    return formula &&
+      'compute' in formula &&
+      typeof formula.compute === 'function'
+      ? (formula as T)
       : undefined
-  return formula &&
-    'compute' in formula &&
-    typeof formula.compute === 'function'
-    ? (formula as ContractFormula<any, any>)
-    : undefined
-}
+  }
 
-export const getWalletFormula = (
-  formulaName: string
-): WalletFormula<any, any> | undefined => {
-  const formulaPath = formulaName.split('/')
-  const formulaBase = formulaPath
-    .slice(0, -1)
-    .reduce(
-      (acc, key) =>
-        typeof acc === 'object' && key in acc
-          ? (acc as NestedFormulaMap<WalletFormula<any, any>>)[key]
-          : undefined,
-      walletFormulas as
-        | NestedFormulaMap<WalletFormula<any, any>>
-        | WalletFormula<any, any>
-        | undefined
-    )
-
-  const formula =
-    typeof formulaBase === 'object'
-      ? (formulaBase as NestedFormulaMap<WalletFormula<any, any>>)[
-          formulaPath[formulaPath.length - 1]
-        ]
-      : undefined
-  return formula &&
-    'compute' in formula &&
-    typeof formula.compute === 'function'
-    ? (formula as WalletFormula<any, any>)
-    : undefined
-}
-
-export const getGenericFormula = (
-  formulaName: string
-): GenericFormula<any, any> | undefined => {
-  const formulaPath = formulaName.split('/')
-  const formulaBase = formulaPath
-    .slice(0, -1)
-    .reduce(
-      (acc, key) =>
-        typeof acc === 'object' && key in acc
-          ? (acc as NestedFormulaMap<GenericFormula<any, any>>)[key]
-          : undefined,
-      genericFormulas as
-        | NestedFormulaMap<GenericFormula<any, any>>
-        | GenericFormula<any, any>
-        | undefined
-    )
-
-  const formula =
-    typeof formulaBase === 'object'
-      ? (formulaBase as NestedFormulaMap<GenericFormula<any, any>>)[
-          formulaPath[formulaPath.length - 1]
-        ]
-      : undefined
-  return formula &&
-    'compute' in formula &&
-    typeof formula.compute === 'function'
-    ? (formula as GenericFormula<any, any>)
-    : undefined
-}
+export const getContractFormula =
+  makeGetFormula<ContractFormula>(contractFormulas)
+export const getGenericFormula = makeGetFormula<GenericFormula>(genericFormulas)
+export const getValidatorFormula =
+  makeGetFormula<ValidatorFormula>(validatorFormulas)
+export const getWalletFormula = makeGetFormula<WalletFormula>(walletFormulas)
 
 export const getTypedFormula = (
   type: FormulaType,
@@ -109,17 +60,24 @@ export const getTypedFormula = (
           type,
           formula: getContractFormula(formulaName),
         }
+      : type === FormulaType.Generic
+      ? {
+          type,
+          formula: getGenericFormula(formulaName),
+        }
+      : type === FormulaType.Validator
+      ? {
+          type,
+          formula: getValidatorFormula(formulaName),
+        }
       : type === FormulaType.Wallet
       ? {
           type,
           formula: getWalletFormula(formulaName),
         }
-      : {
-          type,
-          formula: getGenericFormula(formulaName),
-        }
+      : undefined
 
-  if (!typeAndFormula.formula) {
+  if (!typeAndFormula?.formula) {
     throw new Error(`Formula not found: ${formulaName}`)
   }
