+++ src/test/mocks.ts
@@ -6,11 +6,16 @@ const mocks = {
   },
 }
 
-Object.entries(mocks).forEach(([module, mocks]) => {
-  jest.mock(module, () => mocks)
+// Creates mocks with default implementations.
+export const restoreOriginalMocks = () => {
+  Object.entries(mocks).forEach(([module, mocks]) => {
+    jest.mock(module, () => mocks)
 
-  // Mock with default implementation unless overridden.
-  Object.entries(mocks).forEach(([name, mock]) => {
-    mock.mockImplementation(jest.requireActual(module)[name])
+    // Mock with default implementation unless overridden.
+    Object.entries(mocks).forEach(([name, mock]) => {
+      mock.mockImplementation(jest.requireActual(module)[name])
+    })
   })
-})
+}
+
+restoreOriginalMocks()
